# openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/settings": {
      "get": {
        "operationId": "SettingsAdminController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsListDto_Output"
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/settings/{key}": {
      "put": {
        "operationId": "SettingsAdminController_set",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSettingDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "SettingsAdminController_reset",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/tours/{idOrSlug}": {
      "get": {
        "operationId": "ToursController_get",
        "parameters": [
          {
            "name": "X-Bundle-Id",
            "in": "header",
            "description": "Bundle id of the city app",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idOrSlug",
            "required": true,
            "in": "path",
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourContentDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Missing or unknown X-Bundle-Id"
          }
        },
        "tags": [
          "app"
        ]
      }
    },
...
    "/v1/favorites": {
      "get": {
        "operationId": "FavoritesController_list",
        "parameters": [
          {
            "name": "X-Bundle-Id",
            "in": "header",
            "description": "Bundle id of the city app",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoritesDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Missing or unknown X-Bundle-Id"
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "user": []
          }
        ],
        "tags": [
          "app"
        ]
      }
    },
    "/v1/favorites/tours/{id}": {
      "put": {
        "operationId": "FavoritesController_addTour",
        "parameters": [
          {
            "name": "X-Bundle-Id",
            "in": "header",
            "description": "Bundle id of the city app",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Missing or unknown X-Bundle-Id"
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "user": []
          }
        ],
        "tags": [
          "app"
        ]
      },
      "delete": {
        "operationId": "FavoritesController_removeTour",
        "parameters": [
          {
            "name": "X-Bundle-Id",
            "in": "header",
            "description": "Bundle id of the city app",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Missing or unknown X-Bundle-Id"
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "user": []
          }
        ],
        "tags": [
          "app"
        ]
      }
    },
    "/v1/favorites/points/{id}": {
      "put": {
        "operationId": "FavoritesController_addPoint",
        "parameters": [
          {
            "name": "X-Bundle-Id",
            "in": "header",
            "description": "Bundle id of the city app",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Missing or unknown X-Bundle-Id"
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "user": []
          }
        ],
        "tags": [
          "app"
        ]
      },
      "delete": {
        "operationId": "FavoritesController_removePoint",
        "parameters": [
          {
            "name": "X-Bundle-Id",
            "in": "header",
            "description": "Bundle id of the city app",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Missing or unknown X-Bundle-Id"
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "user": []
          }
        ],
        "tags": [
          "app"
        ]
      }
    },
...
    "/v1/admin/logs/files": {
      "get": {
        "operationId": "LogsAdminController_files",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogFileListDto_Output"
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/logs/files/{name}": {
      "get": {
        "operationId": "LogsAdminController_download",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Log file, one JSON object per line"
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/logs/entries": {
      "get": {
        "operationId": "LogsAdminController_entries",
        "parameters": [
          {
            "name": "file",
            "required": true,
            "in": "query",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "minLevel",
            "required": false,
            "in": "query",
            "schema": {
              "default": "info",
              "type": "string",
              "enum": [
                "fatal",
                "error",
                "warn",
                "info",
                "debug",
                "trace"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 2000,
              "default": 200,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogEntriesDto_Output"
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/logs/level": {
      "get": {
        "operationId": "LogsAdminController_level",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogLevelDto_Output"
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "LogsAdminController_setLevel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogLevelDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    }
  },
  "info": {
    "title": "Vitago API",
    "description": "",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "user": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "admin": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SettingsListDto_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "public": {
                  "type": "boolean"
                },
                "value": {},
                "defaultValue": {},
                "overridden": {
                  "type": "boolean"
                },
                "updatedAt": {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              },
              "required": [
                "key",
                "description",
                "public",
                "value",
                "defaultValue",
                "overridden",
                "updatedAt"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "items"
        ],
        "additionalProperties": false
      },
      "SetSettingDto": {
        "type": "object",
        "properties": {
          "value": {}
        },
        "required": [
          "value"
        ]
      },
...
      "TourContentDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "subtitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "coverImageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "priceKopecks": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "distanceMeters": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "durationMinutes": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "pointCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "introAudioUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mapViewport": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "center": {
                    "type": "object",
                    "properties": {
                      "latitude": {
                        "type": "number",
                        "minimum": -90,
                        "maximum": 90
                      },
                      "longitude": {
                        "type": "number",
                        "minimum": -180,
                        "maximum": 180
                      }
                    },
                    "required": [
                      "latitude",
                      "longitude"
                    ],
                    "additionalProperties": false
                  },
                  "zoom": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 22
                  },
                  "bounds": {
                    "type": "object",
                    "properties": {
                      "south": {
                        "type": "number",
                        "minimum": -90,
                        "maximum": 90
                      },
                      "west": {
                        "type": "number",
                        "minimum": -180,
                        "maximum": 180
                      },
                      "north": {
                        "type": "number",
                        "minimum": -90,
                        "maximum": 90
                      },
                      "east": {
                        "type": "number",
                        "minimum": -180,
                        "maximum": 180
                      }
                    },
                    "required": [
                      "south",
                      "west",
                      "north",
                      "east"
                    ],
                    "additionalProperties": false
                  }
                },
                "required": [
                  "center",
                  "zoom"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ]
          },
          "route": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "const": "LineString"
                  },
                  "coordinates": {
                    "minItems": 2,
                    "type": "array",
                    "items": {
                      "type": "array",
                      "prefixItems": [
                        {
                          "type": "number",
                          "minimum": -180,
                          "maximum": 180
                        },
                        {
                          "type": "number",
                          "minimum": -90,
                          "maximum": 90
                        }
                      ],
                      "items": false,
                      "minItems": 2,
                      "maxItems": 2
                    }
                  }
                },
                "required": [
                  "type",
                  "coordinates"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ]
          },
          "points": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "position": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "latitude": {
                  "type": "number"
                },
                "longitude": {
                  "type": "number"
                },
                "isFree": {
                  "type": "boolean"
                },
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "address": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "openingHours": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "imageUrl": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "markerImageUrl": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "lockedMarkerImageUrl": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "categoryIds": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  }
                },
                "audio": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "url": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "autoplayRadiusMeters": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "durationSeconds": {
                          "anyOf": [
                            {
                              "type": "integer",
                              "minimum": -9007199254740991,
                              "maximum": 9007199254740991
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "transcript": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "subtitles": {
                          "anyOf": [
                            {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "startMs": {
                                    "type": "number"
                                  },
                                  "endMs": {
                                    "type": "number"
                                  },
                                  "text": {
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "startMs",
                                  "endMs",
                                  "text"
                                ],
                                "additionalProperties": false
                              }
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      },
                      "required": [
                        "url",
                        "autoplayRadiusMeters",
                        "durationSeconds",
                        "transcript",
                        "subtitles"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              },
              "required": [
                "id",
                "position",
                "latitude",
                "longitude",
                "isFree",
                "name",
                "description",
                "address",
                "openingHours",
                "imageUrl",
                "markerImageUrl",
                "lockedMarkerImageUrl",
                "categoryIds",
                "audio"
              ],
              "additionalProperties": false
            }
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "slug": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "iconImageUrl": {
                  "type": [
                    "string",
                    "null"
                  ]
                }
              },
              "required": [
                "id",
                "slug",
                "name",
                "iconImageUrl"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "subtitle",
          "summary",
          "coverImageUrl",
          "priceKopecks",
          "distanceMeters",
          "durationMinutes",
          "pointCount",
          "description",
          "introAudioUrl",
          "imageUrls",
          "mapViewport",
          "route",
          "points",
          "categories"
        ],
        "additionalProperties": false
      },
...
    }
  }
}